Extract Reusable ExperimentStatus Component
