+++ pages/experiments/[id].tsx
@@ -1,7 +1,7 @@
 import Button from '@material-ui/core/Button'
 import Grid from '@material-ui/core/Grid'
 import Paper from '@material-ui/core/Paper'
-import { makeStyles, useTheme } from '@material-ui/core/styles'
+import { useTheme } from '@material-ui/core/styles'
 import Table from '@material-ui/core/Table'
 import TableBody from '@material-ui/core/TableBody'
 import TableCell, { TableCellProps } from '@material-ui/core/TableCell'
@@ -9,7 +9,6 @@ import TableHead from '@material-ui/core/TableHead'
 import TableRow from '@material-ui/core/TableRow'
 import Typography from '@material-ui/core/Typography'
 import useMediaQuery from '@material-ui/core/useMediaQuery'
-import clsx from 'clsx'
 import debugFactory from 'debug'
 import { useRouter } from 'next/router'
 import { toIntOrNull } from 'qc-to_int'
@@ -19,6 +18,7 @@ import ExperimentsApi from '@/api/ExperimentsApi'
 import MetricsApi from '@/api/MetricsApi'
 import SegmentsApi from '@/api/SegmentsApi'
 import DatetimeText from '@/components/DatetimeText'
+import ExperimentStatus from '@/components/ExperimentStatus'
 import ExperimentTabs from '@/components/ExperimentTabs'
 import Layout from '@/components/Layout'
 import {
@@ -277,34 +277,6 @@ function VariationsTable(props: { variations: Variation[] }) {
   )
 }
 
-const useStatusStyles = makeStyles({
-  root: {
-    borderRadius: 1,
-    padding: '0.5rem',
-  },
-  completed: {
-    background: '#4caf5014',
-    color: '#4caf50',
-  },
-  disabled: {
-    background: '#82828214',
-    color: '#828282',
-  },
-  running: {
-    background: '#ff980014',
-    color: '#ff9800',
-  },
-  staging: {
-    background: '#82828214',
-    color: '#828282',
-  },
-})
-
-function ExperimentStatus({ experiment }: { experiment: ExperimentFull }) {
-  const classes = useStatusStyles()
-  return <span className={clsx(classes.root, classes[experiment.status])}>{experiment.status}</span>
-}
-
 function ExperimentDetails({
   debugMode,
   experiment,
@@ -325,7 +297,7 @@ function ExperimentDetails({
         <Grid item xs={12}>
           <div>
             <span>{experiment.name}</span>
-            <ExperimentStatus experiment={experiment} />
+            <ExperimentStatus status={experiment.status} />
             <Button variant='contained'>Edit</Button>
           </div>
         </Grid>
